requirements.txt
# pip install -r requirements.txt
fastapi
sqlalchemy
psycopg2
ipython
python-decouple
requests
alembic